Implemented The Ingot Custom Item With Mod And Added Code-Base Support
Implemented The Ingot Custom Item With Mod And Added Code-Base Support. Still need to fix a few bugs with the item, of note being that when assigned material properties the weight of the item goes to 0 for whatever reason, will work on fixing that next, the actually implementing the use of said ingots in the repair service.