Changed Door Bashing AI So More Enemies Can Do It, Such As Bears, 1/25/2021
Changed Door Bashing AI So More Enemies Can Do It, Such As Bears, 1/25/2021. Now on this subject, I may as well also improve the roll formula for door bashing, so the basher's stats or other attributes come into play, not just a flat 20% - lock difficulty, which can go as low as 1%, boring and lame.