Added Basic Implementation Of AI Drinking Potions In Their Inventory, Just Health Potions For Now, Also Door Bashing Based On Strength, 1/26/2021
Added Basic Implementation Of AI Drinking Potions In Their Inventory, Just Health Potions For Now, Also Door Bashing Based On Strength, 1/26/2021. Note about the potion thing, there seems to be some bug that the caster of the original effect does not get recognized after loading a save, which does not seem to matter much for the effect since it mostly just would determine the effect based on their level, which my added potions have no level scaling, but still annoying. This might be a DFU bug and not just exclusive to my enemy potion drinking implementation, not going to worry about that for now, not something I fully understand nor should worry to much about, outside of loading a save the potion drinking and at least the healing part seem to work fine, tomorrow will finish up the other effects that can be used as well as removing the used potions from the enemy inventory upon use, etc.