Added Ingredients To Rarity Based Drop Chance System, Basic Framework Atleast, 2/6/2021
Added Ingredients To Rarity Based Drop Chance System, Basic Framework Atleast, 2/6/2021. So I still need to do a lot of work, especially now on the actual "loot tables" for the various enemies and dungeons in this regard, as well as the rarity and likely gold values of the ingredients and such. But I can now use this framework to make it all work better and hopefully more cleanly than it currently does.