Added A Whole Bunch Of New Drug Effects In The Form Of Custom Diseases For More Control, 2/21/2021
Added A Whole Bunch Of New Drug Effects In The Form Of Custom Diseases For More Control, 2/21/2021. Still need to add the new drugs as items and such, but the effects exist, also need more code changes for this to all fully work, but got the harder part done in this part I would say. Next i'll likely work on the custom sprites and sprite replacements.