Got Most Basic Functionality Of Summon Effect Working, 6/18/2021
Got Most Basic Functionality Of Summon Effect Working. Rat spawns where crosshair is pointing and its Unity object is destroyed when the spell duration ends. Many bugs that need to be fixed though, like the obvious one being if you summon again while the previous is still active, the first summoned one becomes a permanent entity and the latest gets attached to the new effect. Have to make it so the first summoned thing "sticks" to the spell effect until it dies or the spell runs out etc, not bad start though, 6/18/2021.